+++ Web/Contracts/Skipass/Requests/CreateSkipassRequest.cs
@@ -2,7 +2,7 @@ namespace Web.Contracts.Skipass.Requests;
 
 public sealed class CreateSkipassRequest
 {
-    public CreateSkipassRequest(int balance, Guid tariffId, Guid visitorId, bool status)
+    public CreateSkipassRequest(double balance, Guid tariffId, Guid visitorId, bool status)
     {
         Balance = balance;
         TariffId = tariffId;
@@ -10,7 +10,7 @@ public sealed class CreateSkipassRequest
         Status = status;
     }
 
-    public int Balance { get; set; }
+    public double Balance { get; set; }
 
     public Guid TariffId { get; set; }
