+++ Web/Contracts/Tariffication/TarifficationResponse.cs
@@ -2,7 +2,7 @@ namespace Web.Contracts.Tariffication;
 
 public sealed class TarifficationResponse
 {
-    public TarifficationResponse(Guid id, int price, Guid tariffId, Guid locationId)
+    public TarifficationResponse(Guid id, double price, Guid tariffId, Guid locationId)
     {
         Id = id;
         Price = price;
@@ -12,7 +12,7 @@ public sealed class TarifficationResponse
     
     public Guid Id { get; set; }
     
-    public int Price { get; set; }
+    public double Price { get; set; }
     
     public Guid TariffId { get; set; }
